requirements.txt
faker==30.8.1
gocept.pseudonymize==3.0